🗺️ Atlas: [architectural change] Split Distributed Vector Index Blob
